Moved common test helpers to the _suite_test file
